Implement relative selections as store middleware
This is necessary (though possibly not to the best way?) because for
relative selections ("next", "first", "previous"), we need access to the
store (to find out which is, e.g., the "next" note) and we need to
actually load the note (which we don’t want to do in a reducer).

The solution is to check the action before it gets to the reducer and,
while having access to the current, immutable state, to transform it
into another action that gets returned.