fluous: Let map_state_to_props set specific order
Instead of returning a dictionary, we’ll return a tuple of tuples now.
This makes it possible to set properties on GObjects in the specified
order.

For example: when the state contains a "current_id" and a list of
objects this current_id refers to, and you want to add an item, you
first need to updates the list of objects and the current_id afterwards.