Make sure signals are emitted after all properties are set
This makes sense in some in some situations where the GObject needs to
reference properties in the current state. E. g., when changing the
position (x, y), both x and y should be set before the signal handlers
are called because a "move" operation that is invoked by the signal
handler doesn't work when one of the properties is unset.