Add "development mode"
This is important for running two instances (development and production)
at the same time on the same machine.

Otherwise the two apps will try to access the same DBus, which currently
leads to a crash of both instances (which should not happen, but it's
impossible anyway to run the same Gtk program in two versions twice).